verifreg tests
